v9.22.0: Fix Decimal type conversion in PredictiveAnalyzer
- Fixed type error when multiplying Decimal revenue by float growth_rate
- Convert Decimal to float explicitly before arithmetic operations
- Ensures revenue predictions work correctly with PostgreSQL Decimal types

+++ agent_executor.py
@@ -2150,7 +2150,7 @@ class PredictiveAnalyzerAgent(BaseAgent):
 
             # Simple projection (would use ML in production)
             if historical:
-                avg_monthly = sum(h['revenue'] or 0 for h in historical) / len(historical)
+                avg_monthly = float(sum(float(h['revenue'] or 0) for h in historical) / len(historical))
                 growth_rate = 1.1  # 10% growth assumption
 
                 predictions = []
